sqldb: exclude sqlite from the JS and unsupported platform builds
