+++ sqldb/sqlerrors_no_sqlite.go
@@ -1,4 +1,4 @@
-//go:build !js && !(windows && (arm || 386)) && !(linux && (ppc64 || mips || mipsle || mips64))
+//go:build js || (windows && (arm || 386)) || (linux && (ppc64 || mips || mipsle || mips64))
 
 package sqldb
 
@@ -8,8 +8,6 @@ import (
 
 	"github.com/jackc/pgconn"
 	"github.com/jackc/pgerrcode"
-	"modernc.org/sqlite"
-	sqlite3 "modernc.org/sqlite/lib"
 )
 
 var (
@@ -21,12 +19,6 @@ var (
 // MapSQLError attempts to interpret a given error as a database agnostic SQL
 // error.
 func MapSQLError(err error) error {
-	// Attempt to interpret the error as a sqlite error.
-	var sqliteErr *sqlite.Error
-	if errors.As(err, &sqliteErr) {
-		return parseSqliteError(sqliteErr)
-	}
-
 	// Attempt to interpret the error as a postgres error.
 	var pqErr *pgconn.PgError
 	if errors.As(err, &pqErr) {
@@ -36,32 +28,6 @@ func MapSQLError(err error) error {
 	// Return original error if it could not be classified as a database
 	// specific error.
 	return err
-}
-
-// parsePostgresError attempts to parse a sqlite error as a database agnostic
-// SQL error.
-func parseSqliteError(sqliteErr *sqlite.Error) error {
-	switch sqliteErr.Code() {
-	// Handle unique constraint violation error.
-	case sqlite3.SQLITE_CONSTRAINT_UNIQUE:
-		return &ErrSQLUniqueConstraintViolation{
-			DBError: sqliteErr,
-		}
-
-	case sqlite3.SQLITE_CONSTRAINT_PRIMARYKEY:
-		return &ErrSQLUniqueConstraintViolation{
-			DBError: sqliteErr,
-		}
-
-	// Database is currently busy, so we'll need to try again.
-	case sqlite3.SQLITE_BUSY:
-		return &ErrSerializationError{
-			DBError: sqliteErr,
-		}
-
-	default:
-		return fmt.Errorf("unknown sqlite error: %w", sqliteErr)
-	}
 }
 
 // parsePostgresError attempts to parse a postgres error as a database agnostic
